Quote separation, alongside moving half of bot.py to ViviaTools
